requirements.txt
Flask==0.12.1
pip
mongo